refactor(Repeater): Split Phaser/BuildTreeScheduler/CustomProperty per winui-port
Splits the small infrastructure types per the winui-port file layout. Bodies are
1:1 with WinUI commit 4b206bce3. BuildTreeScheduler explicitly Start()s after
Stopwatch.Reset() (Uno-specific note); CustomProperty preserves the C++ field naming.

Tracks issue #22908.
