+++ src/Uno.UI/UI/Xaml/Controls/Repeater/CustomProperty.h.mux.cs
@@ -2,10 +2,14 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 // MUX Reference CustomProperty.h, commit 4b206bce3
 
-using Microsoft.UI.Xaml.Data;
+using System;
 
 namespace Microsoft.UI.Xaml.Controls;
 
-internal partial class CustomProperty : ICustomProperty
+partial class CustomProperty
 {
+	private string m_name;
+	private Type m_typeName;
+	private Func<object, object> m_getter;
+	private Action<object, object> m_setter;
 }
